requirements.txt
pyserial
python-osc
tk
